check if directory exist before patching

+++ app/script/patch-modules.sh
@@ -1,15 +1,22 @@
 #!/bin/bash
-pushd node_modules/adbkit
-rm src 2> /dev/null
-ln -s lib src 2> /dev/null
-popd
 
-pushd node_modules/adbkit-logcat
-rm src 2> /dev/null
-ln -s lib src 2> /dev/null
-popd
+if [ -d "node_modules/adbkit" ]; then
+  pushd node_modules/adbkit
+  rm src 2> /dev/null
+  ln -s lib src 2> /dev/null
+  popd
+fi
 
-pushd node_modules/adbkit-monkey
-rm src 2> /dev/null
-ln -s lib src 2> /dev/null
-popd
+if [ -d "node_modules/adbkit-logcat" ]; then
+  pushd node_modules/adbkit-logcat
+  rm src 2> /dev/null
+  ln -s lib src 2> /dev/null
+  popd
+fi
+
+if [ -d "node_modules/adbkit-monkey" ]; then
+  pushd node_modules/adbkit-monkey
+  rm src 2> /dev/null
+  ln -s lib src 2> /dev/null
+  popd
+fi
